# Vectorize lovasz_softmax loss