#### Changes
Currently the lovasz_softmax loss performs a for loop in range of number of classes. 
This PR removes the for loop.

Speed benchmark (M1 CPU):

```
from kornia.losses import lovasz_softmax_loss
import torch
label = torch.randint(0, 15, (2, 256, 256))
pred = torch.rand(2, 15, 256, 256)
%timeit l=lovasz_softmax_loss(pred, label)
```

Before: 

```
93.8 ms ± 630 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After the change:

```
36.8 ms ± 4.33 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

#### Type of change
- [x] 🔬 New feature (non-breaking change which adds functionality)


#### Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings